refactor(genericidp): dont pass Activity to ViewModel
