+++ auth/app/src/main/java/com/google/firebase/quickstart/auth/kotlin/GenericIdpViewModel.kt
@@ -1,15 +1,12 @@
 package com.google.firebase.quickstart.auth.kotlin
 
-import android.app.Activity
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.auth.ktx.auth
-import com.google.firebase.auth.ktx.oAuthProvider
 import com.google.firebase.ktx.Firebase
-import java.util.ArrayList
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
@@ -25,8 +22,7 @@ class GenericIdpViewModel(
     data class UiState(
         var status: String = "",
         var detail: String? = null,
-        var isSignInVisible: Boolean = true,
-        val providerNames: List<String> = ArrayList(PROVIDER_MAP.keys)
+        var isSignInVisible: Boolean = true
     )
 
     fun showSignedInUser() {
@@ -51,27 +47,8 @@ class GenericIdpViewModel(
         }
     }
 
-    fun signIn(activity: Activity, providerName: String) {
-        // Could add custom scopes here
-        val customScopes = listOf<String>()
-
-        // Examples of provider ID: apple.com (Apple), microsoft.com (Microsoft), yahoo.com (Yahoo)
-        val providerId = PROVIDER_MAP[providerName]!!
-
-        val oAuthProvider = oAuthProvider(providerId) {
-            scopes = customScopes
-        }
-
-        viewModelScope.launch {
-            try {
-                val authResult = firebaseAuth.startActivityForSignInWithProvider(activity, oAuthProvider).await()
-                Log.d(TAG, "activitySignIn:onSuccess:${authResult.user}")
-                updateUiState(authResult.user)
-            } catch (e: Exception) {
-                Log.w(TAG, "activitySignIn:onFailure", e)
-                // TODO(thatfiredev): Snackbar Sign in failed, see logs for details.
-            }
-        }
+    fun showSignedInUser(user: FirebaseUser) {
+        updateUiState(user)
     }
 
     fun signOut() {
@@ -101,11 +78,5 @@ class GenericIdpViewModel(
 
     companion object {
         const val TAG = "GenericIdpViewModel"
-        private val PROVIDER_MAP = mapOf(
-            "Apple" to "apple.com",
-            "Microsoft" to "microsoft.com",
-            "Yahoo" to "yahoo.com",
-            "Twitter" to "twitter.com"
-        )
     }
 }
